envia-multa-servico/requirements.txt
aiobotocore==0.11.1
loafer==1.3.2
python-decouple==3.3
requests==2.23.0
sendgrid==6.2.2